+++ libraries/aircraft-simulator/src/tower.cpp
@@ -20,12 +20,12 @@ Tower &Tower::operator=(const Tower &other)
     return *this;
 }
 
-double Tower::calculateTime(const OurVector<3> &current_position)
+long double Tower::calculateTime(const OurVector<3> &current_position)
 {
     const OurVector<3>& b = _position;
     const OurVector<3>& a = current_position;
 
-    double distance = std::sqrt((std::pow((a[0] - b[0]), 2)) + (std::pow((a[1] - b[1]), 2)) +
+    long double distance = sqrtl((std::pow((a[0] - b[0]), 2)) + (std::pow((a[1] - b[1]), 2)) +
                                (std::pow((a[2] - b[2]), 2)));
 
     return distance / k_wave_speed;
