+++ libraries/math/src/equation_solver.cpp
@@ -34,19 +34,19 @@ OurVector<3> EquationSolver::solve(OurVector<EQUATIONS_COUNT>& tdoas)
     auto equation = [=](const OurVector<3>& at, uint8_t tower_i, uint8_t tower_j)
     {
         auto coordinates_delta_i = _towers_coordinates[tower_i] - at;
-        double d_i = 0;
+        long double d_i = 0;
         for (uint8_t i = 0; i < 3; ++i)
         {
             d_i += std::pow(coordinates_delta_i[i], 2);
         }
-        d_i = std::sqrt(d_i);
+        d_i = sqrtl(d_i);
         auto coordinates_delta_j = _towers_coordinates[tower_j] - at;
-        double d_j = 0;
+        long double d_j = 0;
         for (uint8_t i = 0; i < 3; ++i)
         {
             d_j += std::pow(coordinates_delta_j[i], 2);
         }
-        d_j = std::sqrt(d_j);
+        d_j = sqrtl(d_j);
         return std::abs(d_i - d_j);
     };
 
@@ -74,26 +74,26 @@ OurVector<3> EquationSolver::solve(OurVector<EQUATIONS_COUNT>& tdoas)
     return _initial_coordinates;
 }
 
-double EquationSolver::distance(const OurVector<3>& from, const OurVector<3>& to)
+long double EquationSolver::distance(const OurVector<3>& from, const OurVector<3>& to)
 {
-    return std::sqrt(std::pow(from[0] - to[0], 2) + std::pow(from[1] - to[1], 2) + std::pow(from[2] - to[2], 2));
+    return sqrtl(std::pow(from[0] - to[0], 2) + std::pow(from[1] - to[1], 2) + std::pow(from[2] - to[2], 2));
 }
 
 OurVector<3> EquationSolver::getJacobianRow(OurVector<3>& coordinate, uint8_t tower_i, uint8_t tower_j)
 {
     OurVector<3> jacobian_row;
-    auto numerator = [](double tower_coordinate, double plane_coordinate) { return plane_coordinate - tower_coordinate; };
-    auto denominator = [=](uint8_t index, double x, double y, double z)
+    auto numerator = [](long double tower_coordinate, long double plane_coordinate) { return plane_coordinate - tower_coordinate; };
+    auto denominator = [=](uint8_t index, long double x, long double y, long double z)
     {
-        return std::sqrt(
+        return sqrtl(
             std::pow(_towers_coordinates[index][0] - x, 2) +
             std::pow(_towers_coordinates[index][1] - y, 2) +
             std::pow(_towers_coordinates[index][2] - z, 2)
         );
     };
     
-    double denominator_i = denominator(tower_i, coordinate[0], coordinate[1], coordinate[2]);
-    double denominator_j = denominator(tower_j, coordinate[0], coordinate[1], coordinate[2]);
+    long double denominator_i = denominator(tower_i, coordinate[0], coordinate[1], coordinate[2]);
+    long double denominator_j = denominator(tower_j, coordinate[0], coordinate[1], coordinate[2]);
     assert(denominator_i && denominator_j);
 
     for (int column = 0; column < 3; ++column)
